Enable continuous streaming from Track.Audio
The previous implementation of `Audio` for a track would return a
snapshot up to the end of the audio available at that time.

This will return a Streamer that continues to receive audio chunks as
they become available.

Span.Audio will still take a limited-size snapshot of the audio for
its specified range.